move register and login into userController
