+++ routes/userRoutes.js
@@ -1,8 +1,9 @@
 const express = require("express");
 const router = express.Router();
 const userController = require("../controllers/userController");
+const authenticateJWT = require("../middleware/authMiddleware");
 
-router.get("/", userController.getAllUsers);
-router.post("/", userController.createUser);
+router.post('/register', userController.register);
+router.post('/login', userController.login);
 
 module.exports = router;
